geekbattle/requirements.txt
Django==1.7.2
argparse==1.2.1
wsgiref==0.1.2
twython==3.2.0
facebook-sdk==0.4.0
mechanize
